third_party/WebKit/LayoutTests/imported/wpt/service-workers/service-worker/postmessage-to-client.https-expected.txt

This is a testharness.js-based test.
FAIL postMessage from ServiceWorker to Client assert_equals: message events should use MessageEvent interface. expected function "function MessageEvent() { [native code] }" but got function "function ServiceWorkerMessageEvent() { [native code] }"
Harness: the test ran to completion.
